Refactor Multihash API and make it a struct
This changes the Multihash API to something that's probably more useful.
You can encode a Multihash by instantiating one with the hash
name + digest, and then calling `.to_s` on the instance. You can decode
a Multihash by instantiating it with the fully-encoded digest string and
retrieving the `@len`, `@hash_code`, or `@digest` values.

It's not necessary to require the `len` value to be passed into a
constructor, so I removed that parameter. The length can be found
directly from the length of the digest string. In fact, if we allow it
to be passed in explicitly, then we're inviting inconsistent state.

This adds comparison operators (equals and not-equals), since comparing
cryptographic hashes is a common use case. It makes sense to allow
people to compare Multihashes directly, without the need to explicitly
decode them. This does not override any other operators.

Also, instead of allowing Multihash to be a mutable class, it is now an
immutable struct. Allowing instance variables to change would likely
produce an invalid state.

This also adds some example usage toward the bottom of the module. This
is not meant to stay for good, but it'll be a good start for someone to
take it out and base tests off it. Until we write tests, I think we
should keep this code here; the crystal compiler is strange in that it
sometimes lets errors slip through the cracks unless that code is
actively executed.

Fixes #7, Fixes #9